+++ src/vml/math_1_3op/max2min.sh
@@ -1,0 +1,2 @@
+cat ./vmax.cpp | sed -e 's/max/min/g'  > vmin.cpp
+cat ./mmax.cpp | sed -e 's/max/min/g'  > mmin.cpp